requirements/base.txt
Django==1.10.2
Pillow==3.4.2
django-admin-bootstrapped==2.5.7
django-authtools==1.5.0
django-braces==1.10.0
django-common-helpers==0.9.0
django-crispy-forms==1.6.1
django-environ==0.4.0
django-webpack-loader==0.4.1
easy-thumbnails==2.3
requests==2.12.3
six==1.10.0
slacker==0.9.30
wsgiref==0.1.2
